requirements.txt
torch>=2.2.0
torchvision>=0.17.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
albumentations>=1.3.0
opencv-python>=4.8.0
